+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ### 3.2.56
     - Fixed sync, so it can handle larger data sets.
+    - Rework on inviting to make more reliable.
+    - Invite: Added delay for messages to comply with Blizzard's API.
+    - Rework analytics to more accurately track stats.
 ### 3.2.55
     - Fixed issue with no messages.
     - Fixed database defaults for new installs.